feat: add reusable LanguageSwitcherComponent and integrate into App
