+++ src/app/shared/components/language-switcher/language-switcher.component.html
@@ -1,0 +1,11 @@
+<div class="lang-switcher">
+  <mat-button-toggle-group
+    appearance="standard"
+    class="lang-switcher"
+    name="lang"
+    [value]="translate.currentLang"
+    (change)="switchLanguage($event.value)">
+    <mat-button-toggle value="en">EN</mat-button-toggle>
+    <mat-button-toggle value="pl">PL</mat-button-toggle>
+  </mat-button-toggle-group>
+</div>
